pyproject.toml
[tool.poetry]
name = "spider"
version = "0.1.0"
description = ""
authors = ["Arthur Borges <arthurborges.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
asyncio = "^3.4.3"
python-dotenv = "^0.17.1"
playwright = "^1.12.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
